fix: Adding support for toolchains (#36)
Firstly thanks for creating this project, I am really enjoying using it
🎉

I find myself immediately configuring [biome](https://biomejs.dev/)
after starting a new project. This PR makes it available as an option
(but defaults to `none` so as not to impose it). I do think most users
who intend to build something with tsrouter-app will inevitably look to
install linting/formatting solutions.

I had also considered adding `"eslint/prettier"` (together as a single
option) to the list of toolchains next but held off in case this is not
something you see as the responsibility of this project.

Also a minor fix where the `tailwind` prompt is being skipped when
running `pnpm start` without config (`normalizeOptions` sets it to
false, but the check is currently `cliOptions.tailwind === undefined`).
I have gone through the test steps in the contributing guidelines
without error.